* Makes changes so WithLogger is only accessible for 1.21 and above
