+++ apps/internal/logger/logger_test.go
@@ -1,3 +1,5 @@
+//go:build go1.21
+
 package logger
 
 import (
